#338: DiaryTest 네이밍 Diary로 변경 및 Diary 디렉토리로 파일 위치 변경(DiaryTest 디렉토리 제거)
