+++ Todoary-iOS/Todoary/Data/DataManager/Diary/DiaryDataManager.swift
@@ -13,7 +13,7 @@ class DiaryDataManager {
     let headers : HTTPHeaders = [.authorization(UserDefaults.standard.string(forKey: "accessToken")!)]
     
     //다이어리 생성/수정
-    func posts(viewController: DiaryTestViewController, createdDate: String, parameter: DiaryInput){
+    func posts(viewController: DiaryViewController, createdDate: String, parameter: DiaryInput){
         
         AF.request("https://todoary.com/diary/\(createdDate)", method: .post, parameters: parameter, encoder: JSONParameterEncoder.default, headers: self.headers).validate().responseDecodable(of: ApiModel.self) { response in
             switch response.result {
@@ -68,7 +68,7 @@ class DiaryDataManager {
     }
     
     //다이어리 스티커 생성/수정/삭제
-    func diaryStickerDataManager(viewController: DiaryTestViewController, createdDate: String, parameter: DiaryStickerInput){
+    func diaryStickerDataManager(viewController: DiaryViewController, createdDate: String, parameter: DiaryStickerInput){
         
         AF.request("https://todoary.com/diary/\(createdDate)/sticker", method: .put, parameters: parameter, encoder: JSONParameterEncoder.default, headers: self.headers).validate().responseDecodable(of: DiaryStickerModel.self) { response in
             
@@ -91,7 +91,7 @@ class DiaryDataManager {
     }
     
     
-    func getDiarySticker(viewController: DiaryTestViewController, createdDate: String){
+    func getDiarySticker(viewController: DiaryViewController, createdDate: String){
         
         AF.request("https://todoary.com/diary/\(createdDate)/sticker", method: .get, parameters: nil,  headers: self.headers, interceptor: Interceptor()).validate().responseDecodable(of: GetDiaryStickerModel.self) { response in
             switch response.result {
