--- Todoary-iOS/Todoary/Presentation/Views/Diary/DiaryTest/DiaryStickerTestView.swift
@@ -1,20 +1,0 @@
-//
-//  DiaryStickerTestView.swift
-//  Todoary
-//
-//  Created by 박지윤 on 2022/12/30.
-//
-
-import UIKit
-
-class DiaryStickerTestView: UIView {
-
-    /*
-    // Only override draw() if you perform custom drawing.
-    // An empty implementation adversely affects performance during animation.
-    override func draw(_ rect: CGRect) {
-        // Drawing code
-    }
-    */
-
-}
